requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pymysql
cryptography
python-multipart
passlib
pyjwt
bcrypt==3.2.2
mysql-connector-python